+++ includes/class-tmsm-aquatonic-course-booking.php
@@ -231,15 +231,22 @@ class Tmsm_Aquatonic_Course_Booking {
 
 		$this->loader->add_filter( 'gform_replace_merge_tags', $plugin_public, 'gform_replace_merge_tags_booking', 20, 7 );
 
+		// Ajax frontend
 		$this->loader->add_action( 'wp_ajax_tmsm-aquatonic-course-booking-times', $plugin_public, 'ajax_times' );
 		$this->loader->add_action( 'wp_ajax_nopriv_tmsm-aquatonic-course-booking-times', $plugin_public, 'ajax_times' );
 
+		// Barcode
 		$this->loader->add_action( 'wp_ajax_tmsm-aquatonic-course-booking-generate-barcode', $plugin_public, 'generate_barcode_image', 10 );
 		$this->loader->add_action( 'wp_ajax_nopriv_tmsm-aquatonic-course-booking-generate-barcode', $plugin_public, 'generate_barcode_image', 10 );
 
+		// Emails
 		$this->loader->add_filter( 'gform_html_message_template_pre_send_email', $plugin_public, 'gform_html_message_template_pre_send_email', 20, 1 );
 		$this->loader->add_filter( 'gform_pre_send_email', $plugin_public, 'gform_pre_send_email', 20, 4 );
 		$this->loader->add_filter( 'woocommerce_email_styles', $plugin_public, 'woocommerce_email_styles', 20, 2 );
+
+
+		// Misc
+		$this->loader->add_filter( 'body_class', $plugin_public, 'body_class', 10, 2 );
 
 	}