+++ public/class-tmsm-aquatonic-course-booking-public.php
@@ -1913,4 +1913,28 @@ class Tmsm_Aquatonic_Course_Booking_Public {
 		return $css;
 	}
 
+
+	/**
+	 * Filters the list of CSS body class names
+	 *
+	 * @since 2.8.0
+	 *
+	 * @param string[] $classes An array of body class names.
+	 * @param string[] $class   An array of additional class names added to the body.
+	 *
+	 * @return string[]
+	 */
+	public function body_class( $classes, $class ) {
+		global $post;
+
+		$cancel_page_id = $this->get_option( 'page_cancel_id' );
+		$add_page_id    = $this->get_option( 'page_add_id' );
+
+		if ( ! empty( $post ) && ! empty( $post->ID ) && in_array( $post->ID, [ $cancel_page_id, $add_page_id ] ) ) {
+			$classes[] = 'tmsm-aquatonic-course-booking-pages';
+		}
+
+		return $classes;
+	}
+
 }